+++ src/Contracts/EntityPersisterInterface.php
@@ -16,42 +16,32 @@ interface EntityPersisterInterface
     /**
      * Create a new record returning the hydrated result
      *
-     * Properties is an array of key => value pairs, or nested arrays of scalar data
-     * to send via a POST request to an API end point. For complex data types or
-     * file uploads, custom handling of resources will be necessary.
-     *
      * All exceptions should be converted to the EntityPersisterException type and
      * include an appropriate ClientExceptionInterface trace exception that includes
      * the API response object.
      *
-     * @param array $properties
+     * @param ApiActionInterface $action
      *
      * @return object
      * @throws EntityPersisterException
      */
-    public function store(array $properties): object;
+    public function create(ApiActionInterface $action): object;
 
     /**
-     * Update an existing record with identity $id
-     *
-     * Properties is an array of the values to change. This could be the new object
-     * representation depending on the API implementation. Similar to store, complex
-     * types require specific handling. Updates should make either PUT or PATCH requests
-     * depending on the API implementation.
+     * Update an existing record, returning the new representation
      *
      * All exceptions should be converted to the EntityPersisterException type and
      * include an appropriate ClientExceptionInterface trace exception that includes
      * the API response object.
      *
-     * @param int|string $id
-     * @param array      $properties
+     * @param ApiActionInterface $action
      *
      * @return object
      */
-    public function update($id, array $properties): object;
+    public function update(ApiActionInterface $action): object;
 
     /**
-     * Destroy an existing record with identity $id
+     * Destroy an existing record
      *
      * For the given identity, send a DELETE request to the API and handle the
      * response. True should be returned if successfully removed, or an exception
@@ -61,9 +51,9 @@ interface EntityPersisterInterface
      * include an appropriate ClientExceptionInterface trace exception that includes
      * the API response object.
      *
-     * @param int|string $id
+     * @param ApiActionInterface $action
      *
      * @return bool
      */
-    public function destroy($id): bool;
+    public function destroy(ApiActionInterface $action): bool;
 }